# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Informe Centro-Oeste - News Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/news:
    get:
      operationId: getPublicNews
      tags: [public]
      summary: List published news
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, most_viewed]
            default: latest
      responses:
        "200":
          description: Paginated news list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNews"
  /public/news/{slug}:
    get:
      operationId: getPublicNewsArticle
      tags: [public]
      summary: Get single published article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article with related news
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /public/banners:
    get:
      operationId: getPublicBanners
      tags: [public]
      summary: Get active banners
      parameters:
        - name: position
          in: query
          schema:
            type: string
            enum: [TOP, BELOW_PLAYER, SIDEBAR, BETWEEN_SECTIONS, FOOTER]
      responses:
        "200":
          description: Banners list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
...
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin-categories]
      summary: List categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: adminCreateCategory
      tags: [admin-categories]
      summary: Create category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin-categories]
      summary: Update category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}/status:
    patch:
      operationId: adminToggleCategoryStatus
      tags: [admin-categories]
      summary: Toggle category status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
              required: [isActive]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  # ─── Admin: News ──────────────────────────────────────────────────
  /admin/news:
    get:
      operationId: adminListNews
      tags: [admin-news]
      summary: List all news (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated news
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAdminNews"
    post:
      operationId: adminCreateNews
      tags: [admin-news]
      summary: Create news article
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNewsDetail"
  /admin/news/{id}:
    get:
      operationId: adminGetNews
      tags: [admin-news]
      summary: Get news article (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNewsDetail"
    put:
      operationId: adminUpdateNews
      tags: [admin-news]
      summary: Update news article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNewsDetail"
    delete:
      operationId: adminDeleteNews
      tags: [admin-news]
      summary: Delete news article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/news/{id}/publish:
    patch:
      operationId: adminPublishNews
      tags: [admin-news]
      summary: Publish news article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNewsDetail"
  /admin/news/{id}/archive:
    patch:
      operationId: adminArchiveNews
      tags: [admin-news]
      summary: Archive news article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNewsDetail"

  # ─── Admin: Banners ───────────────────────────────────────────────
  /admin/banners:
    get:
      operationId: adminListBanners
      tags: [admin-banners]
      summary: List banners
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: adminCreateBanner
      tags: [admin-banners]
      summary: Create banner
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /admin/banners/{id}:
    put:
      operationId: adminUpdateBanner
      tags: [admin-banners]
      summary: Update banner
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: adminDeleteBanner
      tags: [admin-banners]
      summary: Delete banner
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/banners/{id}/status:
    patch:
      operationId: adminToggleBannerStatus
      tags: [admin-banners]
      summary: Toggle banner status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
              required: [isActive]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"

  # ─── Admin: Videos ────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, isActive, createdAt]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
      required: [name, slug]
    NewsCard:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
          nullable: true
        featuredImage:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        viewCount:
          type: integer
        isFeatured:
          type: boolean
        category:
          $ref: "#/components/schemas/Category"
        author:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required: [id, name]
      required: [id, title, slug, viewCount, isFeatured, category, author]
    ArticleDetail:
      allOf:
        - $ref: "#/components/schemas/NewsCard"
        - type: object
          properties:
            content:
              type: string
            seoTitle:
              type: string
              nullable: true
            seoDescription:
              type: string
              nullable: true
            seoKeywords:
              type: string
              nullable: true
            tags:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  slug:
                    type: string
                required: [id, name, slug]
            related:
              type: array
              items:
                $ref: "#/components/schemas/NewsCard"
          required: [content]
    PaginatedNews:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NewsCard"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    AdminNewsDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
          nullable: true
        content:
          type: string
        featuredImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        categoryId:
          type: integer
        authorId:
          type: integer
        isFeatured:
          type: boolean
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        seoKeywords:
          type: string
          nullable: true
        viewCount:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        author:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required: [id, name]
        tags:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              slug:
                type: string
            required: [id, name, slug]
      required: [id, title, slug, content, status, categoryId, authorId, isFeatured, viewCount, createdAt, updatedAt, category, author, tags]
    CreateNewsRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        content:
          type: string
        featuredImage:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          default: DRAFT
        categoryId:
          type: integer
        isFeatured:
          type: boolean
          default: false
        seoTitle:
          type: string
        seoDescription:
          type: string
        seoKeywords:
          type: string
        tagIds:
          type: array
          items:
            type: integer
      required: [title, slug, content, categoryId]
    PaginatedAdminNews:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminNewsDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        position:
          type: string
          enum: [TOP, BELOW_PLAYER, SIDEBAR, BETWEEN_SECTIONS, FOOTER]
        imageUrl:
          type: string
        targetUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, position, imageUrl, isActive, createdAt]
    CreateBannerRequest:
      type: object
      properties:
        title:
          type: string
        position:
          type: string
          enum: [TOP, BELOW_PLAYER, SIDEBAR, BETWEEN_SECTIONS, FOOTER]
        imageUrl:
          type: string
        targetUrl:
          type: string
        isActive:
          type: boolean
          default: true
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
      required: [title, position, imageUrl]
...