# lib/api-zod/src/generated/types/newsCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Informe Centro-Oeste - News Portal API
 * OpenAPI spec version: 1.0.0
 */
import type { Category } from "./category";
import type { NewsCardAuthor } from "./newsCardAuthor";

export interface NewsCard {
  id: number;
  title: string;
  slug: string;
  summary?: string | null;
  featuredImage?: string | null;
  publishedAt?: Date | null;
  viewCount: number;
  isFeatured: boolean;
  category: Category;
  author: NewsCardAuthor;
}